+++ pyccel/ast/class_defs.py
@@ -6,7 +6,7 @@
 This module contains all types which define a python class which is automatically recognised by pyccel
 """
 
-from pyccel.ast.builtin_methods.list_methods import ListAppend, ListInsert, ListPop
+from pyccel.ast.builtin_methods.list_methods import ListAppend, ListInsert, ListPop, ListClear
 
 
 from .builtins  import PythonImag, PythonReal, PythonConjugate
@@ -142,6 +142,8 @@ ListClass = ClassDef('list', class_type = NativeHomogeneousList(),
             PyccelFunctionDef('insert', func_class = ListInsert,
                 decorators = {}),
             PyccelFunctionDef('pop', func_class = ListPop,
+                decorators = {}),
+            PyccelFunctionDef('clear', func_class = ListClear,
                 decorators = {}),
         ])